requirements.txt
gptpdf
requests
fastapi
uvicorn
python-dotenv
pydantic
aiofiles
